Add files via upload
Dynamic Pool Simulation
Pool Simulation

Description
This GitHub project presents a dynamic simulation of a pool that allows users to visualize and interact with three different simulation models. The simulation accurately represents the movement and rotation of pool balls on a pool table.

Features
Position Showing Model: The first simulation model displays the positions of the pool balls on the pool table as they move and collide with each other. Users can observe the realistic motion of the balls across the table.

Position and Rotation in Pool Plane Model: The second simulation model enhances the visualization by not only showing the positions of the pool balls but also illustrating their rotations within the pool plane. This enables users to get a better understanding of the rotational dynamics during the ball movements.

Position and Rotation in All Directions Model: The third and most advanced simulation model extends the visualization beyond the pool plane. Users can now observe the positions and rotations of the pool balls in all directions, providing a more immersive experience and better insights into the three-dimensional dynamics.